+++ tests/serde_helpers/mod.rs
@@ -15,13 +15,20 @@ where
     // Log XML that we try to deserialize to see it in the failed tests output
     dbg!(source);
     let mut de = Deserializer::from_str(source);
+    assert_eq!(
+        de.resolver().level(),
+        0,
+        "no user namespace bindings expected just after creation: {:#?}",
+        de.resolver()
+    );
+
     let result = T::deserialize(&mut de);
 
     // If type was deserialized, the whole XML document should be consumed
     if result.is_ok() {
         de.check_eof_reached();
 
-        let resolver = de.get_ref().get_ref().resolver();
+        let resolver = de.resolver();
         assert_eq!(
             resolver.level(),
             0,